requirements.txt
python-telegram-bot==20.8
APScheduler
motor
pymongo
httpx
python-dotenv
uvicorn
starlette
nanoid
